move stuff out of blif::reader and into submodules
