+++ vts_core/src/blif/mod.rs
@@ -1,5 +1,11 @@
 pub mod command;
+pub mod error;
 pub mod netlist;
 pub mod reader;
 
+// TODO(rikus): Should these be made `pub`?
+mod buffer;
+mod token;
+
+pub use error::Error;
 pub use reader::BlifReader;
